requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==6.0.1  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
rcssmin==1.1.0  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.1.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
flower==1.0.0  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.2.12  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.48.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis

# Wagtail
# ------------------------------------------------------------------------------
wagtail==2.16.2  # https://github.com/wagtail/wagtail


# Wagtail Recaptcha
# ------------------------------------------------------------------------------
django-recaptcha==2.0.6
wagtail-django-recaptcha==1.0

# Wagtail Menu
# ------------------------------------------------------------------------------
wagtailmenus==3.0.2

# Wagtail Localize
# ------------------------------------------------------------------------------
wagtail-localize==1.2.1